# TPBookstore-api/config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
...
  /api/v1/order:
    post:
      summary: Create new order
      description: Create new order
      tags:
        - Order
      security:
        - bearerAuth: []
      requestBody:
        description: Order form
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderForm'
      responses:
        '400':
          description: No order items/One or more product order quantity exceed available quantity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Admin get orders
      description: Admin get orders
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: dateOrder
          required: false 
          schema:
            type: string
          description: Sort orders by date (newest/latest)
        - in: query
          name: status
          required: false
          schema: 
            type: string
          description: Filter orders by status (disabled/notDisabled/all)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/order/{id}:
    get:
      summary: Get order by id
      description: Get order by id
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an order
      description: Delete an order
      security:
        - bearerAuth: []
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/category: 
    post: 
      description: Create a new category
      summary: Create a new category
      tags:
        - Category
      security:
        - bearerAuth: []
      requestBody:
        description: Payload to create a new category
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryForm'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary:  Get all categories (non-user, user, admin)
      description: Get all categories (non-user, user, admin)
      tags:
        - Category
      parameters:
        - in: query
          name: dateOrder
          required: false 
          schema:
            type: string
          description: Sort categories by date (newest/latest)
        - in: query
          name: status
          required: false
          schema: 
            type: string
          description: Filter categories by status (disabled/notDisabled/all) 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  # /api/v1/category/all:
  #   get:
  #     summary:  Admin get all categories 
  #     description: Admin get all categories 
  #     tags:
  #       - Category
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: '#/components/schemas/Category' 
  /api/v1/category/{id}: 
    put:
      summary: Update exsited category
      tags:
        - Category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: string
      requestBody:
        description: Payload to update a new category
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryForm'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'   
    delete:
      summary: Delete existed category
      description: Delete existed category
      security:
        - bearerAuth: []
      tags:
        - Category
      parameters:
        - name: id
          in: path
          description: Category ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
...
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          description: The date and time when this record is created
          format: date-time
        updatedAt:
          type: string
          description: The date and time when this record is updated
          format: date-time
...
    OrderItem:
      type: object
      properties:
        name:
          type: string
          description: Order item name
        qty:
          type: number
          description: Order item quantity
        image:
          type: string
          description: Order item image's url
        price:
          type: number
          description: Order item price
        product:
          type: string
          description: Order item ID
    ShippingAddress:
      type: object
      properties:
        address:
          type: string
          description: Shipping address
        city:
          type: string
          description: Shipping city
        postalCode:
          type: string
          description: Shipping postal code
        country:
          type: string
          description: Shipping country
...
    OrderForm:
      type: object
      properties:
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          description: Payment method
          default: Paypal
        itemsPrice:
          type: number
          description: Items price
        taxPrice:
          type: number
          description: Tax price
          default: 0
        shippingPrice:
          type: number
          description: Shipping price
          default: 0
        totalPrice:
          type: number
          description: Total price
          default: 0
    Order:
      type: object
      properties:
        user:
          type: string
          description: Ordered user ID
        orderItems:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/OrderItem'
              - type: object
                properties:
                  _id:
                    type: string
                    description: Order item ID
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          description: Payment method
          default: Paypal
        taxPrice:
          type: number
          description: Tax price
          default: 0
        shippingPrice:
          type: number
          description: Shipping price
          default: 0
        totalPrice:
          type: number
          description: Total price
          default: 0     
        isPaid:
          type: boolean
          description: Is paid yet ?
        isDelivered:
          type: boolean
          description: Is deliverd yet ?
        _id: 
          type: string
          description: Order ID
        timestamp: 
          $ref: '#/components/schemas/Timestamp'
        __v: 
          type: number
          description: Document version key
        isDisabled:
          type: boolean
          description: is Disabled or not
...
    CategoryForm:
      type: object
      properties:
        name:
          type: string
          description: Category name
        slug:
          type: string
          description: Category slug
        status:
          type: boolean
          description: Category status
          default: true 
    Category:
      type: object
      properties:
        _id: 
          type: string
          description: Category Id
        name:
          type: string
          description: Category name
        slug:
          type: string
          description: Category slug
        status:
          type: boolean
          description: Category status
          default: true    
        timestamp:
          $ref: '#/components/schemas/Timestamp' 
        createdBy:
          type: string
          description: User object id who created this record
        updatedBy:
          type: string
          description: User object id who updated this record
        isDisabled:
          type: boolean
          description: is Disabled or not
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above